% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_scripts.R
\name{getDfFromPySpDir}
\alias{getDfFromPySpDir}
\title{Create R dataframe from the Presient 1.0 Populator output (reads the pysp format)}
\usage{
getDfFromPySpDir(fullPath)
}
\arguments{
\item{fullPath}{The path to the directory full of scenarios and actuals.}
}
\value{
Returns a dataframe of the actuals, forecast, and all scenarios with their probabilities.
}
\description{
Create R dataframe from the Presient 1.0 Populator output (reads the pysp format)
}
